Make Pool automatically do a PING periodically
In the past I've said I didn't want to do this because it adds an extra
go-routine to the Pool that it doesn't technically _have_ to have, and
also makes the Empty method's behavior a bit ambiguous. Previously Empty
would merely run through the pool buffer and close/discard all
connections. Now it will also stop that go-routine. Using the Pool after
calling Empty was previously fine, and technically it still is, but now
the pinging won't happen. I have a feeling this will cause some
confusion for someone at some point, but using Pool after Empty is such
a strange thing to do it's probably fine.